Added extra statement to compare character encoding
For some reason, in readData() when the EGSE sends uplink or MDAQ packets, it doesn't like the encoding. Upon debugging it appears that the character is already decoded. 

This could be due to the ground station equipment, or it could be something to do with how the program reads in null chars